Time: 8 ms (44.94%), Space: 38.8 MB (11.55%) - LeetHub

+++ 1708-design-parking-system/1708-design-parking-system.cpp
@@ -1,6 +1,6 @@
 class ParkingSystem {
 private:
-    int cars[4];
+    short cars[4];
 public:
     ParkingSystem(int big, int medium, int small) {
         cars[1] = big;
